Add file overriding to allow the use of custom LTF and FXX files when importing sounds

+++ Program.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Diagnostics;
 using System.IO;
+using System.Text;
 
 namespace H3EK_FaceFX_Wrapper
 {
@@ -12,6 +13,39 @@ namespace H3EK_FaceFX_Wrapper
             // generate an LTF with FaceFX and convert it to FXX for tool
             if (args[0] == "-exec")
             {
+                bool FXXOverride = false;
+                bool LTFOverride = false;
+                string fxxOverridePath = "";
+                string ltfOverridePath = "";
+                // read config
+                if (File.Exists("FaceFXWrapper.cfg"))
+                {
+                    using (StreamReader reader = new StreamReader(File.OpenRead("FaceFXWrapper.cfg"), Encoding.UTF8))
+                    {
+                        string line = reader.ReadLine();
+                        while (line != null)
+                        {
+                            if (line.StartsWith("FXXOverridePath"))
+                            {
+                                int from = line.IndexOf("\"") + 1;
+                                int to = line.LastIndexOf("\"");
+                                fxxOverridePath = line.Substring(from, to - from);
+                                if (fxxOverridePath != "")
+                                    FXXOverride = true;
+                            }
+                            else if (line.StartsWith("LTFOverridePath"))
+                            {
+                                int from = line.IndexOf("\"") + 1;
+                                int to = line.LastIndexOf("\"");
+                                ltfOverridePath = line.Substring(from, to - from);
+                                if (ltfOverridePath != "")
+                                    LTFOverride = true;
+                            }
+                            line = reader.ReadLine();
+                        }
+                    }
+                }
+
                 string commands = args[1];
 
                 // split FaceFX command args
@@ -40,30 +74,42 @@ namespace H3EK_FaceFX_Wrapper
                 pFrom = exportanimArgs.IndexOf("\"") + "\"".Length;
                 string ltfPath = exportanimArgs.Substring(pFrom, exportanimArgs.Length - pFrom - 1);
 
-                // write commands to file to use with FaceFX's exec arg
-                using (StreamWriter sw = File.CreateText("commands.fxl"))
+                // handle overrides
+                if (FXXOverride)
+                    File.Copy(fxxOverridePath, ltfPath.Replace(".ltf", ".fxx"));
+                else if (LTFOverride)
+                    File.Copy(ltfOverridePath, ltfPath);
+
+                if (!LTFOverride && !FXXOverride)
                 {
-                    sw.WriteLine("robobrad -load default"); // DAO FxStudio will crash if robobrad isn't loaded
-                    sw.WriteLine("loadActor " + loadactorArgs);
-                    sw.WriteLine("analyze " + analyzeArgs);
-                    sw.WriteLine("exportLTF " + exportanimArgs);
-                    sw.WriteLine("// LTF path: " + ltfPath);
+                    // write commands to file to use with FaceFX's exec arg
+                    using (StreamWriter sw = File.CreateText("commands.fxl"))
+                    {
+                        sw.WriteLine("robobrad -load default"); // DAO FxStudio will crash if robobrad isn't loaded
+                        sw.WriteLine("loadActor " + loadactorArgs);
+                        sw.WriteLine("analyze " + analyzeArgs);
+                        sw.WriteLine("exportLTF " + exportanimArgs);
+                        sw.WriteLine("// LTF path: " + ltfPath);
+                    }
+
+                    // launch FaceFX
+                    Process fxStudio = new Process();
+                    fxStudio.StartInfo.FileName = "FxStudioOriginal.exe";
+                    fxStudio.StartInfo.Arguments = $"-exec commands.fxl";
+                    fxStudio.Start();
+
+                    // wait for FaceFX to produce our LTF file
+                    while (!File.Exists(ltfPath)) ;
                 }
 
-                // launch FaceFX
-                Process fxStudio = new Process();
-                fxStudio.StartInfo.FileName = "FxStudioOriginal.exe";
-                fxStudio.StartInfo.Arguments = $"-exec commands.fxl";
-                fxStudio.Start();
-
-                // wait for FaceFX to produce our LTF file
-                while (!File.Exists(ltfPath)) ;
-
-                // load and convert the LTF file
-                FXX_File FXX = new FXX_File();
-                if (FXX.loadLTF(ltfPath)) // write FXX if the LTF was loaded successfully
-                    FXX.WriteTo(ltfPath.Replace(".ltf", ".fxx"));
-                File.Delete(ltfPath);
+                if (!FXXOverride)
+                {
+                    // load and convert the LTF file
+                    FXX_File FXX = new FXX_File();
+                    if (FXX.loadLTF(ltfPath)) // write FXX if the LTF was loaded successfully
+                        FXX.WriteTo(ltfPath.Replace(".ltf", ".fxx"));
+                    File.Delete(ltfPath);
+                }
             }
             // convert an LTF file directly to FXX
             else if (args[0] == "-convertLTF")
